requirements-production.txt
# =============================================================================
# RefurbAdmin AI - Production Requirements
# =============================================================================
# South African Context: POPIA Compliant
# Python 3.12+
# =============================================================================

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# -----------------------------------------------------------------------------
# Authentication & Security
# -----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# -----------------------------------------------------------------------------
# Caching & Rate Limiting
# -----------------------------------------------------------------------------
redis==5.0.1
hiredis==2.3.2

# -----------------------------------------------------------------------------
# Email & Communication
# -----------------------------------------------------------------------------
# Email service
# WhatsApp (optional - uncomment if needed)
# twilio==8.11.1

# -----------------------------------------------------------------------------
# Reports & Documents
# -----------------------------------------------------------------------------
reportlab==4.0.8
jinja2==3.1.3
openpyxl==3.1.2
pandas==2.1.4

# -----------------------------------------------------------------------------
# Monitoring & Observability
# -----------------------------------------------------------------------------
prometheus-client==0.19.0
structlog==24.1.0

# -----------------------------------------------------------------------------
# Testing (Production includes test capabilities)
# -----------------------------------------------------------------------------
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dateutil==2.8.2
pytz==2023.3
httpx==0.26.0
aiofiles==23.2.1

# -----------------------------------------------------------------------------
# South African Specific
# -----------------------------------------------------------------------------
# Phone number validation
phonenumbers==8.13.27

# -----------------------------------------------------------------------------
# Performance
# -----------------------------------------------------------------------------
orjson==3.9.12
uvloop==0.19.0; sys_platform != 'win32'

# -----------------------------------------------------------------------------
# Security Scanning
# -----------------------------------------------------------------------------
bandit==1.7.5
safety==2.3.5
